+++ udata/core/dataservices/rdf.py
@@ -1,13 +1,14 @@
 
 from datetime import datetime
 from typing import List, Optional
-from rdflib import RDF, Graph, URIRef
+from rdflib import RDF, BNode, Graph, Literal, URIRef
 
 from udata.core.dataservices.models import Dataservice, HarvestMetadata as HarvestDataserviceMetadata
 from udata.core.dataset.models import Dataset, License
-from udata.core.dataset.rdf import sanitize_html
+from udata.core.dataset.rdf import dataset_to_graph_id, sanitize_html
 from udata.harvest.models import HarvestSource
-from udata.rdf import DCAT, DCT, contact_point_from_rdf, rdf_value, remote_url_from_rdf, theme_labels_from_rdf, themes_from_rdf, url_from_rdf
+from udata.rdf import DCATAP, TAG_TO_EU_HVD_CATEGORIES, namespace_manager, DCAT, DCT, contact_point_from_rdf, rdf_value, remote_url_from_rdf, themes_from_rdf, url_from_rdf
+from udata.uris import endpoint_for
 
 def dataservice_from_rdf(graph: Graph, dataservice: Dataservice, node, all_datasets: List[Dataset]) -> Dataservice :
     '''
@@ -56,3 +57,53 @@ def dataservice_from_rdf(graph: Graph, dataservice: Dataservice, node, all_datas
     dataservice.tags = themes_from_rdf(d)
 
     return dataservice
+
+
+def dataservice_to_rdf(dataservice, graph=None):
+    '''
+    Map a dataservice domain model to a DCAT/RDF graph
+    '''
+    # Use the unlocalized permalink to the dataset as URI when available
+    # unless there is already an upstream URI
+    if dataservice.harvest and dataservice.harvest.rdf_node_id_as_url:
+        id = URIRef(dataservice.harvest.rdf_node_id_as_url)
+    elif dataservice.id:
+        id = URIRef(endpoint_for('dataservices.show_redirect', 'api.dataservice',
+                    dataservice=dataservice.id, _external=True))
+    else:
+        # Should not happen in production. Some test only
+        # `build()` a dataset without saving it to the DB.
+        id = BNode()
+
+    # Expose upstream identifier if present
+    if dataservice.harvest and dataservice.harvest.dct_identifier:
+        identifier = dataservice.harvest.dct_identifier
+    else:
+        identifier = dataservice.id
+    graph = graph or Graph(namespace_manager=namespace_manager)
+
+    d = graph.resource(id)
+    d.set(RDF.type, DCAT.DataService)
+    d.set(DCT.identifier, Literal(identifier))
+    d.set(DCT.title, Literal(dataservice.title))
+    d.set(DCT.description, Literal(dataservice.description))
+    d.set(DCT.issued, Literal(dataservice.created_at))
+    
+    if dataservice.base_api_url:
+        d.set(DCAT.endpointURL, Literal(dataservice.base_api_url))
+    
+    if dataservice.endpoint_description_url:
+        d.set(DCAT.endpointDescription, Literal(dataservice.endpoint_description_url))
+
+    for tag in dataservice.tags:
+        d.add(DCAT.keyword, Literal(tag))
+
+    # `dataset_to_graph_id(dataset)` URIRef may not exist in the current page
+    # but should exists in the catalog somewhere. Maybe we should create a Node 
+    # with some basic information about this dataset (but this will return a page 
+    # with more datasets than the page size… and could be problematic when processing the
+    # correct Node with all the information in a future page)
+    for dataset in dataservice.datasets:
+        d.add(DCAT.servesDataset, dataset_to_graph_id(dataset))
+
+    return d