+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Update markdown base settings [#3067](https://github.com/opendatateam/udata/pull/3067)
 - Prevent tagging as spam owners' messages [#3071](https://github.com/opendatateam/udata/pull/3071)
 - Add api endpoint /me/org_topics/ [#3070](https://github.com/opendatateam/udata/pull/3070)
+- Expose dataservices in RDF catalog [#3058](https://github.com/opendatateam/udata/pull/3058)
 
 ## 9.0.0 (2024-06-07)